requirements.txt
requests
beautifulsoup4
line-bot-sdk
slackclient
click